Adapt CMake scripts to emscripten
Delete the haru.cmake file as it's useless: the MATH_LIB variable is never used elsewhere and there's another `find_library` for the math library in the root `CMakeLists.txt`.
Don't try to link the Math library when compiling with emscripten, as it's already provided by the system (https://github.com/emscripten-core/emscripten/issues/2600).
Add details about emscripten and the math library in the CMake summary.